Twitter Sentiment Classifier
This is code that takes in a CSV file of twitter data (the text of a tweet, the number of retweets of that tweet, and the number of replies to that tweet) and detects how positive or negative each tweet is. It then creates a CSV file that contains the Number of Retweets, Number of Replies, Positive Score (which is how many happy words are in the tweet), Negative Score (which is how many angry words are in the tweet), and the Net Score (how positive or negative the text is overall) for each tweet. This file can then be uploaded to any spreadsheet program to produce a visual graph of the information. Aside from the "resulting_data.csv", all files mentioned were provided by the course organizers.